+++ imsdk/src/main/java/com/masonsoft/imsdk/core/IMSessionMessageUploadManager.java
@@ -161,7 +161,7 @@ public class IMSessionMessageUploadManager {
              */
             public float mSendProgress;
 
-            public long mErrorCode;
+            public int mErrorCode;
             public String mErrorMessage;
 
             //////////////////////////////////////////////////////////////////////
